+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v1.7.42.1
+## mm/dd/2023
+
+2. [](#improved)
+   * Quick fix for `isDangerousFunction` when `$name` was a closure [#3727](https://github.com/getgrav/grav/issues/3727)
+
 # v1.7.42
 ## 06/14/2023